metrics: add change counters
